added-non-working-second-stage-bootloader

+++ memory/mem_operations.c
@@ -10,7 +10,7 @@ void memory_copy(char* from, char* to, int num_of_bytes) {
     }
 }
 
-void memory_set(u8 *dest, u8 val, u32 len) {
+void memory_set(u32 *dest, u8 val, u32 len) {
     u8 *temp = (u8 *)dest;
     for ( ; len != 0; len--) *temp++ = val;
 }
