+++ memory/mem_operations.h
@@ -8,6 +8,6 @@
 #include "../cpu/types.h"
 
 void memory_copy(char* from, char* to, int num_of_bytes);
-void memory_set(u8 *dest, u8 val, u32 len);
+void memory_set(u32 *dest, u8 val, u32 len);
 
 #endif //OS_32_MEM_OPERATIONS_H
